Caching for SWAP*, misc performance improvements (#64)
* Simplify local search loop structure

* Add update() facility to LocalSearch

* Pass Params to operators

* Stub for caching SWAP* moves

* RELOCATE* should test both ways

* Rename test() to evaluate()

* Caching in SWAP*, RELOCATE* applies all improving moves

* Simplify SWAP*

* Much more aggressively cut (likely) bad SWAP* moves

* Fix RELOCATE*

* Inline two private member functions, and postpone setting update flag until needed

* Simplify updating interface

* Re-use valid jump nodes when updating routes

* Simplify RELOCATE* a little

* Simplify jump nodes and angle calculations

* Simplify route tw/load/dist calculations further

* Reset rather than recreate in Matrix/SWAP*::init

+++ hgs_vrptw/include/CommandLine.h
@@ -78,7 +78,8 @@ public:
             else if (std::string(argv[i]) == "-repairBooster")
                 config.repairBooster = static_cast<size_t>(atoi(argv[i + 1]));
             else if (std::string(argv[i]) == "-intensificationProbability")
-                config.intensificationProbability = atoi(argv[i + 1]);
+                config.intensificationProbability
+                    = static_cast<size_t>(atoi(argv[i + 1]));
             else if (std::string(argv[i]) == "-diversityWeight")
                 config.diversityWeight = atof(argv[i + 1]);
             else if (std::string(argv[i])
